some scoreboard tweaking
